requirements.txt
# 🔧 Core do Motor Tributário
fastapi==0.115.12
uvicorn==0.34.2
SQLAlchemy==2.0.40
python-dotenv==1.1.0
psycopg2-binary

# 📊 Visualização e Dash
dash==3.0.4
dash-bootstrap-components==2.0.2
plotly==6.0.1

# 🤖 IA e embeddings
sentence-transformers
langchain
openai==1.76.2
scikit-learn==1.6.1

# 🧠 NLP e suporte vetorial
faiss-cpu==1.11.0
tqdm==4.67.1

annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
dash==3.0.4
dash-bootstrap-components==2.0.2
dash-extensions==2.0.4
dataclass-wizard==0.35.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
EditorConfig==0.17.0
email_validator==2.2.0
et_xmlfile==2.0.0
faiss-cpu==1.11.0
fastapi==0.115.12
Flask==3.0.3
Flask-Caching==2.3.1
fonttools==4.57.0
fpdf==1.7.2
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpie==3.2.4
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsbeautifier==1.15.4
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-core==0.3.58
langchain-text-splitters==0.3.8
langsmith==0.3.41
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.7.0
multidict==6.4.3
narwhals==1.37.0
nest-asyncio==1.6.0
numpy==2.2.5
openai==1.76.2
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
plotly==6.0.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rich==14.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.0.6
XlsxWriter==3.2.3
zipp==3.21.0
zstandard==0.23.0
psycopg2-binary
sentence-transformers
# 📅 Processamento de datas
dateparser==1.2.0

